email hook works plus test
